requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
cloudinary==1.32.0
dj-database-url==1.2.0
django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version < '3.10'
markdown==3.4.1
pillow==9.4.0
psycopg2==2.9.5
pyjwt==2.6.0
pytz==2022.7.1
requests==2.28.2
setuptools==67.3.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
zipp==3.14.0 ; python_version >= '3.7'
